deps/v8/test/inspector/regress/regress-crbug-488754138-expected.txt

Regression test for crbug.com/488754138

Running test: repro
